autogen_service/pyproject.toml
[project]
name = "phdapply-autogen-service"
version = "0.1.0"
description = "Microsoft AutoGen backend for PhDApply"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "pypdf>=5.1.0",
  "autogen-agentchat>=0.4.8",
  "autogen-ext[openai]>=0.4.8",
  "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.tools", "app.utils"]
